+++ include/nstd/audio.h
@@ -1,6 +1,7 @@
 #ifndef NSTD_AUDIO_H_INCLUDED
 #define NSTD_AUDIO_H_INCLUDED
 #include "audio/audio.h"
+#include "audio/def.h"
 #include "audio/device.h"
 #include "audio/host.h"
 #include "nstd.h"